Remove ifdefs from TimeZoneInfo
Move Win32 and Unix -specific implementation details to their own files.
